+++ data/types/basal/automated/automated_test.go
@@ -11,6 +11,7 @@ import (
 	"github.com/tidepool-org/platform/data/types/basal/automated"
 	testDataTypesBasalAutomated "github.com/tidepool-org/platform/data/types/basal/automated/test"
 	testDataTypesBasal "github.com/tidepool-org/platform/data/types/basal/test"
+	testDataTypesInsulin "github.com/tidepool-org/platform/data/types/insulin/test"
 	testDataTypes "github.com/tidepool-org/platform/data/types/test"
 	testErrors "github.com/tidepool-org/platform/errors/test"
 	"github.com/tidepool-org/platform/pointer"
@@ -32,6 +33,7 @@ func NewAutomated() *automated.Automated {
 	datum.DeliveryType = "automated"
 	datum.Duration = pointer.Int(test.RandomIntFromRange(automated.DurationMinimum, automated.DurationMaximum))
 	datum.DurationExpected = pointer.Int(test.RandomIntFromRange(*datum.Duration, automated.DurationMaximum))
+	datum.InsulinType = testDataTypesInsulin.NewInsulinType()
 	datum.Rate = pointer.Float64(test.RandomFloat64FromRange(automated.RateMinimum, automated.RateMaximum))
 	datum.ScheduleName = pointer.String(testDataTypesBasal.NewScheduleName())
 	return datum
@@ -45,6 +47,7 @@ func CloneAutomated(datum *automated.Automated) *automated.Automated {
 	clone.Basal = *testDataTypesBasal.CloneBasal(&datum.Basal)
 	clone.Duration = test.CloneInt(datum.Duration)
 	clone.DurationExpected = test.CloneInt(datum.DurationExpected)
+	clone.InsulinType = testDataTypesInsulin.CloneInsulinType(datum.InsulinType)
 	clone.Rate = test.CloneFloat64(datum.Rate)
 	clone.ScheduleName = test.CloneString(datum.ScheduleName)
 	return clone
@@ -79,6 +82,7 @@ var _ = Describe("Automated", func() {
 			Expect(datum.DeliveryType).To(Equal("automated"))
 			Expect(datum.Duration).To(BeNil())
 			Expect(datum.DurationExpected).To(BeNil())
+			Expect(datum.InsulinType).To(BeNil())
 			Expect(datum.Rate).To(BeNil())
 			Expect(datum.ScheduleName).To(BeNil())
 		})
@@ -296,6 +300,19 @@ var _ = Describe("Automated", func() {
 					testErrors.WithPointerSourceAndMeta(structureValidator.ErrorValueNotInRange(604800001, 0, 604800000), "/duration", NewMeta()),
 					testErrors.WithPointerSourceAndMeta(structureValidator.ErrorValueNotInRange(604800001, 0, 604800000), "/expectedDuration", NewMeta()),
 				),
+				Entry("insulin type missing",
+					func(datum *automated.Automated) { datum.InsulinType = nil },
+				),
+				Entry("insulin type invalid",
+					func(datum *automated.Automated) {
+						datum.InsulinType.Formulation = nil
+						datum.InsulinType.Mix = nil
+					},
+					testErrors.WithPointerSourceAndMeta(structureValidator.ErrorValueNotExists(), "/insulinType/formulation", NewMeta()),
+				),
+				Entry("insulin type valid",
+					func(datum *automated.Automated) { datum.InsulinType = testDataTypesInsulin.NewInsulinType() },
+				),
 				Entry("rate missing",
 					func(datum *automated.Automated) { datum.Rate = nil },
 					testErrors.WithPointerSourceAndMeta(structureValidator.ErrorValueNotExists(), "/rate", NewMeta()),
@@ -329,6 +346,8 @@ var _ = Describe("Automated", func() {
 						datum.DeliveryType = "invalidDeliveryType"
 						datum.Duration = nil
 						datum.DurationExpected = pointer.Int(604800001)
+						datum.InsulinType.Formulation = nil
+						datum.InsulinType.Mix = nil
 						datum.Rate = pointer.Float64(100.1)
 						datum.ScheduleName = pointer.String("")
 					},
@@ -336,6 +355,7 @@ var _ = Describe("Automated", func() {
 					testErrors.WithPointerSourceAndMeta(structureValidator.ErrorValueNotEqualTo("invalidDeliveryType", "automated"), "/deliveryType", &basal.Meta{Type: "invalidType", DeliveryType: "invalidDeliveryType"}),
 					testErrors.WithPointerSourceAndMeta(structureValidator.ErrorValueNotExists(), "/duration", &basal.Meta{Type: "invalidType", DeliveryType: "invalidDeliveryType"}),
 					testErrors.WithPointerSourceAndMeta(structureValidator.ErrorValueNotInRange(604800001, 0, 604800000), "/expectedDuration", &basal.Meta{Type: "invalidType", DeliveryType: "invalidDeliveryType"}),
+					testErrors.WithPointerSourceAndMeta(structureValidator.ErrorValueNotExists(), "/insulinType/formulation", &basal.Meta{Type: "invalidType", DeliveryType: "invalidDeliveryType"}),
 					testErrors.WithPointerSourceAndMeta(structureValidator.ErrorValueNotInRange(100.1, 0.0, 100.0), "/rate", &basal.Meta{Type: "invalidType", DeliveryType: "invalidDeliveryType"}),
 					testErrors.WithPointerSourceAndMeta(structureValidator.ErrorValueEmpty(), "/scheduleName", &basal.Meta{Type: "invalidType", DeliveryType: "invalidDeliveryType"}),
 				),
@@ -371,6 +391,9 @@ var _ = Describe("Automated", func() {
 				),
 				Entry("does not modify the datum; duration expected missing",
 					func(datum *automated.Automated) { datum.DurationExpected = nil },
+				),
+				Entry("does not modify the datum; insulin type missing",
+					func(datum *automated.Automated) { datum.InsulinType = nil },
 				),
 				Entry("does not modify the datum; rate missing",
 					func(datum *automated.Automated) { datum.Rate = nil },
@@ -437,6 +460,19 @@ var _ = Describe("Automated", func() {
 				),
 				Entry("annotations valid",
 					func(datum *automated.SuppressedAutomated) { datum.Annotations = testData.NewBlobArray() },
+				),
+				Entry("insulin type missing",
+					func(datum *automated.SuppressedAutomated) { datum.InsulinType = nil },
+				),
+				Entry("insulin type invalid",
+					func(datum *automated.SuppressedAutomated) {
+						datum.InsulinType.Formulation = nil
+						datum.InsulinType.Mix = nil
+					},
+					testErrors.WithPointerSource(structureValidator.ErrorValueNotExists(), "/insulinType/formulation"),
+				),
+				Entry("insulin type valid",
+					func(datum *automated.SuppressedAutomated) { datum.InsulinType = testDataTypesInsulin.NewInsulinType() },
 				),
 				Entry("rate missing",
 					func(datum *automated.SuppressedAutomated) { datum.Rate = nil },
@@ -506,6 +542,9 @@ var _ = Describe("Automated", func() {
 				),
 				Entry("does not modify the datum; annotations missing",
 					func(datum *automated.SuppressedAutomated) { datum.Annotations = nil },
+				),
+				Entry("does not modify the datum; insulin type missing",
+					func(datum *automated.SuppressedAutomated) { datum.InsulinType = nil },
 				),
 				Entry("does not modify the datum; rate missing",
 					func(datum *automated.SuppressedAutomated) { datum.Rate = nil },