+++ data/types/basal/temporary/temporary.go
@@ -4,6 +4,7 @@ import (
 	"github.com/tidepool-org/platform/data"
 	"github.com/tidepool-org/platform/data/types/basal"
 	dataTypesBasalScheduled "github.com/tidepool-org/platform/data/types/basal/scheduled"
+	"github.com/tidepool-org/platform/data/types/insulin"
 	"github.com/tidepool-org/platform/pointer"
 	"github.com/tidepool-org/platform/service"
 	"github.com/tidepool-org/platform/structure"
@@ -30,11 +31,12 @@ type Suppressed interface {
 type Temporary struct {
 	basal.Basal `bson:",inline"`
 
-	Duration         *int       `json:"duration,omitempty" bson:"duration,omitempty"`
-	DurationExpected *int       `json:"expectedDuration,omitempty" bson:"expectedDuration,omitempty"`
-	Percent          *float64   `json:"percent,omitempty" bson:"percent,omitempty"`
-	Rate             *float64   `json:"rate,omitempty" bson:"rate,omitempty"`
-	Suppressed       Suppressed `json:"suppressed,omitempty" bson:"suppressed,omitempty"`
+	Duration         *int                 `json:"duration,omitempty" bson:"duration,omitempty"`
+	DurationExpected *int                 `json:"expectedDuration,omitempty" bson:"expectedDuration,omitempty"`
+	InsulinType      *insulin.InsulinType `json:"insulinType,omitempty" bson:"insulinType,omitempty"`
+	Percent          *float64             `json:"percent,omitempty" bson:"percent,omitempty"`
+	Rate             *float64             `json:"rate,omitempty" bson:"rate,omitempty"`
+	Suppressed       Suppressed           `json:"suppressed,omitempty" bson:"suppressed,omitempty"`
 }
 
 func New() *Temporary {
@@ -50,6 +52,7 @@ func (t *Temporary) Parse(parser data.ObjectParser) error {
 
 	t.Duration = parser.ParseInteger("duration")
 	t.DurationExpected = parser.ParseInteger("expectedDuration")
+	t.InsulinType = insulin.ParseInsulinType(parser.NewChildObjectParser("insulinType"))
 	t.Percent = parser.ParseFloat("percent")
 	t.Rate = parser.ParseFloat("rate")
 	t.Suppressed = parseSuppressed(parser.NewChildObjectParser("suppressed"))
@@ -75,6 +78,9 @@ func (t *Temporary) Validate(validator structure.Validator) {
 	} else {
 		expectedDurationValidator.InRange(DurationMinimum, DurationMaximum)
 	}
+	if t.InsulinType != nil {
+		t.InsulinType.Validate(validator.WithReference("insulinType"))
+	}
 	validator.Float64("percent", t.Percent).InRange(PercentMinimum, PercentMaximum)
 	validator.Float64("rate", t.Rate).Exists().InRange(RateMinimum, RateMaximum)
 	validateSuppressed(validator.WithReference("suppressed"), t.Suppressed)
@@ -87,6 +93,9 @@ func (t *Temporary) Normalize(normalizer data.Normalizer) {
 
 	t.Basal.Normalize(normalizer)
 
+	if t.InsulinType != nil {
+		t.InsulinType.Normalize(normalizer.WithReference("insulinType"))
+	}
 	if t.Suppressed != nil {
 		t.Suppressed.Normalize(normalizer.WithReference("suppressed"))
 	}
@@ -96,10 +105,11 @@ type SuppressedTemporary struct {
 	Type         *string `json:"type,omitempty" bson:"type,omitempty"`
 	DeliveryType *string `json:"deliveryType,omitempty" bson:"deliveryType,omitempty"`
 
-	Annotations *data.BlobArray `json:"annotations,omitempty" bson:"annotations,omitempty"`
-	Percent     *float64        `json:"percent,omitempty" bson:"percent,omitempty"`
-	Rate        *float64        `json:"rate,omitempty" bson:"rate,omitempty"`
-	Suppressed  Suppressed      `json:"suppressed,omitempty" bson:"suppressed,omitempty"`
+	Annotations *data.BlobArray      `json:"annotations,omitempty" bson:"annotations,omitempty"`
+	InsulinType *insulin.InsulinType `json:"insulinType,omitempty" bson:"insulinType,omitempty"`
+	Percent     *float64             `json:"percent,omitempty" bson:"percent,omitempty"`
+	Rate        *float64             `json:"rate,omitempty" bson:"rate,omitempty"`
+	Suppressed  Suppressed           `json:"suppressed,omitempty" bson:"suppressed,omitempty"`
 }
 
 func ParseSuppressedTemporary(parser data.ObjectParser) *SuppressedTemporary {
@@ -124,6 +134,7 @@ func (s *SuppressedTemporary) Parse(parser data.ObjectParser) error {
 	s.DeliveryType = parser.ParseString("deliveryType")
 
 	s.Annotations = data.ParseBlobArray(parser.NewChildArrayParser("annotations"))
+	s.InsulinType = insulin.ParseInsulinType(parser.NewChildObjectParser("insulinType"))
 	s.Percent = parser.ParseFloat("percent")
 	s.Rate = parser.ParseFloat("rate")
 	s.Suppressed = parseSuppressed(parser.NewChildObjectParser("suppressed"))
@@ -138,6 +149,9 @@ func (s *SuppressedTemporary) Validate(validator structure.Validator) {
 	if s.Annotations != nil {
 		s.Annotations.Validate(validator.WithReference("annotations"))
 	}
+	if s.InsulinType != nil {
+		s.InsulinType.Validate(validator.WithReference("insulinType"))
+	}
 	validator.Float64("percent", s.Percent).InRange(PercentMinimum, PercentMaximum)
 	validator.Float64("rate", s.Rate).Exists().InRange(RateMinimum, RateMaximum)
 	validateSuppressed(validator.WithReference("suppressed"), s.Suppressed)
@@ -146,6 +160,9 @@ func (s *SuppressedTemporary) Validate(validator structure.Validator) {
 func (s *SuppressedTemporary) Normalize(normalizer data.Normalizer) {
 	if s.Annotations != nil {
 		s.Annotations.Normalize(normalizer.WithReference("annotations"))
+	}
+	if s.InsulinType != nil {
+		s.InsulinType.Normalize(normalizer.WithReference("insulinType"))
 	}
 	if s.Suppressed != nil {
 		s.Suppressed.Normalize(normalizer.WithReference("suppressed"))