+++ data/types/bolus/test/bolus.go
@@ -2,6 +2,7 @@ package test
 
 import (
 	"github.com/tidepool-org/platform/data/types/bolus"
+	testDataTypesInsulin "github.com/tidepool-org/platform/data/types/insulin/test"
 	testDataTypes "github.com/tidepool-org/platform/data/types/test"
 )
 
@@ -10,6 +11,7 @@ func NewBolus() *bolus.Bolus {
 	datum.Base = *testDataTypes.NewBase()
 	datum.Type = "bolus"
 	datum.SubType = testDataTypes.NewType()
+	datum.InsulinType = testDataTypesInsulin.NewInsulinType()
 	return datum
 }
 
@@ -20,5 +22,6 @@ func CloneBolus(datum *bolus.Bolus) *bolus.Bolus {
 	clone := &bolus.Bolus{}
 	clone.Base = *testDataTypes.CloneBase(&datum.Base)
 	clone.SubType = datum.SubType
+	clone.InsulinType = testDataTypesInsulin.CloneInsulinType(datum.InsulinType)
 	return clone
 }